+++ internal/stremio/addon/client.go
@@ -6,6 +6,8 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"strings"
+	"time"
 
 	"github.com/MunifTanjim/stremthru/core"
 	"github.com/MunifTanjim/stremthru/internal/request"
@@ -14,7 +16,12 @@ import (
 )
 
 var DefaultHTTPTransport = request.DefaultHTTPTransport
-var DefaultHTTPClient = request.DefaultHTTPClient
+var DefaultHTTPClient = func() *http.Client {
+	return &http.Client{
+		Transport: DefaultHTTPTransport,
+		Timeout:   30 * time.Second,
+	}
+}()
 
 type ClientConfig struct {
 	HTTPClient *http.Client
@@ -59,6 +66,13 @@ func (e *ResponseError) Error() string {
 
 func processResponseBody(res *http.Response, err error, v any) error {
 	if err != nil {
+		return err
+	}
+
+	contentType := res.Header.Get("Content-Type")
+	if !strings.Contains(contentType, "application/json") {
+		err := core.NewAPIError("unxpected content-type: " + contentType)
+		err.StatusCode = res.StatusCode
 		return err
 	}